+++ src/senaite/impress/publishview.py
@@ -252,6 +252,13 @@ class PublishView(BrowserView):
             return default
         return orientation
 
+    def store_multireports_individually(self):
+        """Returns the configured setting from the registry
+        """
+        store_individually = api.get_registry_record(
+            "senaite.impress.store_multireports_individually")
+        return store_individually
+
     def get_report_template(self, template=None):
         """Returns the path of report template
         """